% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_metal_currency.R
\docType{package}
\name{cbr}
\alias{cbr}
\alias{cbr-package}
\alias{rus2num}
\title{cbr}
\usage{
rus2num(x)
}
\arguments{
\item{x}{the string with the number}
}
\value{
numeric the number converted from the string
}
\description{
cbr
}
\examples{
rus2num('34 345,34')
}
\author{
Marcel Salikhov, Boris Demeshev
Convert string with a number in Russian tradition in numeric

Convert string with a number in Russian tradition in numeric

Russian standards prescribes to use comma as a decimal separator.
This function removes spaces and converts string to number.
}

