% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_balances}
\alias{cbr_balances}
\title{Credit institutions balances on correspondent accounts with Bank of Russia.}
\usage{
cbr_balances(from = "2007-01-01", to = "2013-01-01")
}
\arguments{
\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with credit institutions balance from cbr.ru
}
\description{
Credit institutions balances on correspondent accounts with Bank of Russia.
}
\details{
Credit institutions balances on correspondent accounts with Bank of Russia.
}
\examples{
df <- cbr_balances(from = '2012-01-01', to = '2013-01-09')
}

