% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_coins}
\alias{cbr_coins}
\title{Release Prices of the Bank of Russia for Investment Coins.}
\usage{
cbr_coins(from = "2007-01-01", to = "2013-01-01")
}
\arguments{
\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical release prices for investment coins from cbr.ru
}
\description{
Release Prices of the Bank of Russia for Investment Coins.
}
\details{
Release Prices of the Bank of Russia for Investment Coins.
}
\examples{
df <- cbr_coins(from = '2012-01-01', to = '2013-01-09')
}

