% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_credit_rates}
\alias{cbr_credit_rates}
\title{Historical credit rates on Interbank Credit Market.}
\usage{
cbr_credit_rates(from = "2007-01-01", to = "2013-01-01")
}
\arguments{
\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical credit rates from cbr.ru
}
\description{
Historical credit rates on Interbank Credit Market.
}
\details{
Historical credit rates on Interbank Credit Market.
}
\examples{
df <- cbr_credit_rates(from = '2012-01-01', to = '2013-01-09')
}

