% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_cur_ondate}
\alias{cbr_cur_ondate}
\title{Actual currency prices from Central Bank of Russia on date.}
\usage{
cbr_cur_ondate(ondate = "2016-06-14")
}
\arguments{
\item{ondate}{the date of actual currency prices, character or Date}
}
\value{
data.frame with actual currency prices on date from cbr.ru
}
\description{
Actual currency prices from Central Bank of Russia on date.
}
\details{
Actual currency prices from Central Bank of Russia on date.
}
\examples{
df <- cbr_cur_ondate(ondate = '2016-06-14')
}

