% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_currency}
\alias{cbr_currency}
\title{Historical currency prices from Central Bank of Russia.}
\usage{
cbr_currency(currency = "R01239", from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{currency}{internal Central Bank currency code starting with the letter 'R' -- full list at https://cbr.ru/scripts/XML_val.asp?d=0}

\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical currency prices from cbr.ru
}
\description{
Historical currency prices from Central Bank of Russia.
}
\examples{
df <- cbr_currency(currency = "R01239", from = "2024-09-01", to = "2024-09-10") # Euro
}
