% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_metal}
\alias{cbr_metal}
\title{Historical metal prices from Central Bank of Russia.}
\usage{
cbr_metal(from = "2007-01-01", to = "2013-01-01")
}
\arguments{
\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical metal prices from cbr.ru
}
\description{
Historical metal prices from Central Bank of Russia.
}
\details{
Historical metal prices from Central Bank of Russia.
}
\examples{
df <- cbr_metal(from = '2012-01-01', to = '2013-01-09')
}

