% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cbr_requests_unsupported}
\alias{cbr_requests_unsupported}
\title{The data.frame of unsupported securities requests}
\format{data.frame with 19 rows and 4 variables
\describe{
  \item{name}{security name}
  \item{node}{short-cut for parsing XML tree}
  \item{type}{type of request structure}
  \item{status}{indicator whether the request is checked}
}}
\source{
manual check
}
\usage{
cbr_requests_unsupported
}
\description{
A data.frame containing the list of securities requests 
for Central Bank of Russia unsupported by this package.
}
\keyword{datasets}

