% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_security_info}
\alias{cbr_security_info}
\title{Convert list from xml to data.frame without changes}
\usage{
cbr_security_info(name, firstDate = as.Date("2000-01-01"),
  secondDate = Sys.Date())
}
\arguments{
\item{name}{the name of variable from http://www.cbr.ru/secinfo/secinfo.asmx}

\item{firstDate}{start of period, Date, as.Date("2000-01-01") by default}

\item{secondDate}{end of period, Date, Sys.Date() by default}
}
\value{
data.frame
}
\description{
Convert list from xml to data.frame without changes
}
\details{
Convert list from xml to data.frame without changes
}
\examples{
df <- cbr_security_info("AuctionsXML")
}

