% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_swap}
\alias{cbr_swap}
\title{Historical swap prices from Central Bank of Russia.}
\usage{
cbr_swap(from = "2007-01-01", to = "2013-01-01")
}
\arguments{
\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical swap prices from cbr.ru
}
\description{
Historical swap prices from Central Bank of Russia.
}
\details{
Historical swap prices from Central Bank of Russia.
}
\examples{
df <- cbr_swap(from = '2012-01-01', to = '2013-01-09')
}

