% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{chr2date}
\alias{chr2date}
\title{Convert character in Date format needed to create url}
\usage{
chr2date(chr)
}
\arguments{
\item{chr}{the date in character format}
}
\value{
Date the date converted from the character
}
\description{
Convert character in Date format needed to create url
}
\details{
Convert character in Date format needed to create url
}
\examples{
chr2date('2016-06-14')
}

