% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swap_metal_currency.R
\name{droplast}
\alias{droplast}
\title{Drops last entry in the list x if this entry is character}
\usage{
droplast(x)
}
\arguments{
\item{x}{the list}
}
\value{
list with omitted last element
}
\description{
Drops last entry in the list x if this entry is character
}
\details{
Drops last entry in the list x if this entry is character
}
\examples{
a <- list(x = 5, y = 'last')
cbr:::droplast(a)
}

