% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{excel2date}
\alias{excel2date}
\title{Convert excel numeric date encoding to date}
\usage{
excel2date(x)
}
\arguments{
\item{x}{the vector of numeric date codes}
}
\value{
the date
}
\description{
Convert excel numeric date encoding to date
}
\details{
While reading excel files dates are sometimes replaced by their numeric codes.
This function recovers original dates from these codes.
}
\examples{
excel2date(12345)
}

