% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{extractXML}
\alias{extractXML}
\title{Convert list from xml to data.frame with little changes}
\usage{
extractXML(url)
}
\arguments{
\item{url}{the url of xml list of data}
}
\value{
data.frame
}
\description{
Convert list from xml to data.frame with little changes
}
\details{
Convert list from xml to data.frame with little changes
}
\examples{
extractXML('http://www.cbr.ru/scripts/xml_swap.asp?date_req1=01/12/2002&date_req2=06/12/2002')
}

