% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr.R
\name{cbr_currency}
\alias{cbr_currency}
\title{Historical currency prices from Central Bank of Russia.}
\usage{
cbr_currency(currency = "R01120", from = "1993-01-05", to = "2013-09-18")
}
\arguments{
\item{currency}{internal Central Bank currency code starting with the letter 'R'}

\item{from}{the first day of the time interval, character or Date}

\item{to}{the last day of the time interval, character or Date}
}
\value{
data.frame with historical currency prices from cbr.ru
}
\description{
Historical currency prices from Central Bank of Russia.
}
\details{
Historical currency prices from Central Bank of Russia.
}
\examples{
df <- cbr_currency(currency = 'R01120',
  from = '1993-01-05', to = '2013-01-09')
# 'R01120' --- Burundi frank :)
}

